feat: separate quotation step eth price
