+++ src/hooks/ramp/useQuote.tsx
@@ -1,11 +1,7 @@
-import { truncateDecimal } from "@/services/truncate"
 import { PaymentMethodType } from "@/types/payment-method.type"
 import { ProviderType } from "@/types/provider.type"
 import { Quote } from "@/types/quote.type"
-import { useEffect } from "react"
 import useSWR from "swr"
-import { quoteVar } from "./useControlModal"
-
 
 export default function useQuoteRamp(
     fiatCurrencyCode: string,
@@ -22,15 +18,8 @@ export default function useQuoteRamp(
     if (toChain && toToken) {
         url += `&toChain=${toChain}&toToken=${toToken}`
     }
-    const { data, error, isLoading, isValidating } = useSWR<Quote>((chainId && fiatCurrencyCode && amount && provider && paymentMethod) ? url : null, { refreshInterval, revalidateOnMount: true })
-    useEffect(() => {
-        if (data) {
-            quoteVar({
-                ...data,
-                amountToken: truncateDecimal(data?.amountToken)
-            })
-        }
-    }, [data])
+    const { data, error, isLoading, isValidating } = useSWR<Quote>((chainId && fiatCurrencyCode && amount && provider && paymentMethod) ? url : null)
+
     return { quote: data, error, isLoading, isValidating }
